wip: move high reward validators task to another place
