+++ src/duty/duty.metrics.ts
@@ -2,7 +2,7 @@ import { LOGGER_PROVIDER } from '@lido-nestjs/logger';
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
 
 import { ConfigService } from 'common/config';
-import { BlockHeaderResponse, ConsensusProviderService } from 'common/eth-providers';
+import { ConsensusProviderService } from 'common/eth-providers';
 import { PrometheusService, TrackTask } from 'common/prometheus';
 
 import { AttestationMetrics } from './attestation';
@@ -40,18 +40,5 @@ export class DutyMetrics {
       this.proposeMetrics.calculate(epoch, possibleHighRewardValidators),
       this.syncMetrics.calculate(epoch, possibleHighRewardValidators),
     ]);
-  }
-
-  @TrackTask('high-reward-validators')
-  public async getPossibleHighRewardValidators(): Promise<string[]> {
-    const actualSlotHeader = <BlockHeaderResponse>await this.clClient.getBlockHeader('head');
-    const headEpoch = BigInt(actualSlotHeader.header.message.slot) / BigInt(this.config.get('FETCH_INTERVAL_SLOTS'));
-    this.logger.log('Getting possible high reward validator indexes');
-    const propDependentRoot = await this.clClient.getDutyDependentRoot(headEpoch);
-    const [sync, prop] = await Promise.all([
-      this.clClient.getSyncCommitteeInfo('finalized', headEpoch),
-      this.clClient.getCanonicalProposerDuties(headEpoch, propDependentRoot),
-    ]);
-    return [...new Set([...prop.map((v) => v.validator_index), ...sync.validators])];
   }
 }
